services/extraction-service/requirements.txt
flask==3.0.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
xlrd==2.0.1
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
chardet==5.2.0
python-magic==0.4.27
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.1
